improve logging
Now that we're subdividing the map into smaller chunks and using
`rayon` to process them in parallel, the current logging is somewhat
less useful. An individual thread may process any number of chunks, so
we no longer have a clear sense of how close to completion the program
is.

This branch reduces the amount of logging in
`AllChargerLocations::find_gaps` significantly --- that method now only
logs when a chunk has completed. In addition, I've added a counter of
how many chunks are completed, so that we can have the main method print
out how close we are to completion, and time how long it took to
process each individual chunk. This should hopefully give us a nice
sense of the program's progress as it executes.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>